+++ cosmwasm/enclaves/shared/contract-engine/src/contract_operations.rs
@@ -1,34 +1,30 @@
-use log::*;
+use cosmos_proto::tx::signing::SignMode;
 
+use cw_types_generic::{BaseAddr, BaseEnv};
+
+use cw_types_v010::encoding::Binary;
+use cw_types_v010::types::CanonicalAddr;
+
+use cw_types_v1::types::{Addr, Timestamp};
+
+use enclave_cosmos_types::types::{ContractCode, HandleType, SigInfo};
+use enclave_crypto::{Ed25519PublicKey, HASH_SIZE};
 use enclave_ffi_types::{Ctx, EnclaveError};
+use enclave_utils::coalesce;
+use log::*;
 
 use crate::contract_validation::{ReplyParams, ValidatedMessage};
 use crate::external::results::{HandleSuccess, InitSuccess, QuerySuccess};
-use crate::wasm::CosmWasmApiVersion;
-use cosmos_proto::tx::signing::SignMode;
-use cw_types_generic::{BaseEnv, CwEnv};
-use cw_types_v010;
-use cw_types_v010::encoding::Binary;
-use cw_types_v010::types::CanonicalAddr;
-use cw_types_v1::addresses::Addr;
-use cw_types_v1::results::{DecryptedReply, Event, Reply, SubMsgResponse, SubMsgResult};
-use cw_types_v1::timestamp::Timestamp;
-use enclave_cosmos_types::types::{ContractCode, HandleType, SigInfo};
-
-use enclave_crypto::{Ed25519PublicKey, HASH_SIZE};
-use enclave_utils::coalesce;
+use crate::message::{parse_message, ParsedMessage};
 
 use super::contract_validation::{
-    extract_contract_key, generate_encryption_key, validate_contract_key, validate_msg,
-    verify_params, ContractKey,
+    generate_encryption_key, validate_contract_key, validate_msg, verify_params, ContractKey,
 };
 use super::gas::WasmCosts;
 use super::io::encrypt_output;
 use super::module_cache::create_module_instance;
 use super::types::{IoNonce, SecretMessage};
 use super::wasm::{ContractInstance, ContractOperation, Engine};
-
-const HEX_ENCODED_HASH_SIZE: usize = HASH_SIZE * 2;
 
 /*
 Each contract is compiled with these functions already implemented in wasm:
@@ -54,37 +50,18 @@ pub fn init(
     msg: &[u8],         // probably function call and args
     sig_info: &[u8],    // info about signature verification
 ) -> Result<InitSuccess, EnclaveError> {
+    trace!("Starting init");
+
     let contract_code = ContractCode::new(contract);
     let contract_hash = contract_code.hash();
 
-    let base_env: BaseEnv = serde_json::from_slice(env).map_err(|err| {
-        warn!(
-            "init got an error while trying to deserialize env input bytes into json {:?}: {}",
-            String::from_utf8_lossy(&env),
-            err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let base_env: BaseEnv = extract_base_env(env)?;
 
     let (sender, contract_address, block_height, sent_funds) = base_env.get_verification_params();
 
-    let canonical_contract_address = CanonicalAddr::from_human(&contract_address).map_err(|err| {
-        warn!(
-            "init got an error while trying to deserialize env_v010.contract.address from bech32 string to bytes {:?}: {}",
-            contract_address, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
-    //
-    // trace!("init env_v010: {:?}", env_v010);
-    //
-    let canonical_sender_address = CanonicalAddr::from_human(&sender).map_err(|err| {
-        warn!(
-            "init got an error while trying to deserialize env_v010.message.sender from bech32 string to bytes {:?}: {}",
-            sender, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let canonical_contract_address = to_canonical(contract_address)?;
+
+    let canonical_sender_address = to_canonical(sender)?;
 
     let contract_key = generate_encryption_key(
         &canonical_sender_address,
@@ -92,18 +69,9 @@ pub fn init(
         &contract_hash,
         &canonical_contract_address,
     )?;
-    trace!("init contract key: {:?}", hex::encode(contract_key));
 
-    let parsed_sig_info: SigInfo = serde_json::from_slice(sig_info).map_err(|err| {
-        warn!(
-            "init got an error while trying to deserialize env input bytes into json {:?}: {}",
-            String::from_utf8_lossy(&sig_info),
-            err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let parsed_sig_info: SigInfo = extract_sig_info(sig_info)?;
 
-    trace!("init input before decryption: {:?}", base64::encode(&msg));
     let secret_msg = SecretMessage::from_slice(msg)?;
 
     verify_params(
@@ -119,12 +87,7 @@ pub fn init(
     let ValidatedMessage {
         validated_msg,
         reply_params,
-    } = validate_msg(&decrypted_msg, contract_code.hash(), None)?;
-
-    trace!(
-        "init input after decryption: {:?}",
-        String::from_utf8_lossy(&validated_msg)
-    );
+    } = validate_msg(&decrypted_msg, &contract_hash, None)?;
 
     let mut engine = start_engine(
         context,
@@ -152,9 +115,6 @@ pub fn init(
     let output = coalesce!(EnclaveError, {
         let vec_ptr = engine.init(env_ptr, msg_info_ptr, msg_ptr)?;
         let output = engine.extract_vector(vec_ptr)?;
-        // TODO: copy cosmwasm's structures to enclave
-        // TODO: ref: https://github.com/CosmWasm/cosmwasm/blob/b971c037a773bf6a5f5d08a88485113d9b9e8e7b/packages/std/src/init_handle.rs#L129
-        // TODO: ref: https://github.com/CosmWasm/cosmwasm/blob/b971c037a773bf6a5f5d08a88485113d9b9e8e7b/packages/std/src/query.rs#L13
         let output = encrypt_output(
             output,
             &secret_msg,
@@ -181,344 +141,14 @@ pub fn init(
     })
 }
 
-pub struct ParsedMessage {
-    pub should_validate_sig_info: bool,
-    pub was_msg_encrypted: bool,
-    pub secret_msg: SecretMessage,
-    pub decrypted_msg: Vec<u8>,
-    pub contract_hash_for_validation: Option<Vec<u8>>,
-}
-
-pub fn redact_custom_events(reply: &mut Reply) {
-    reply.result = match &reply.result {
-        SubMsgResult::Ok(r) => {
-            let mut events: Vec<Event> = Default::default();
-
-            let filtered_attributes = vec!["contract_address".to_string(), "code_id".to_string()];
-            for ev in r.events.iter() {
-                if !ev.ty.starts_with("wasm") {
-                    continue;
-                }
-
-                let mut new_ev = Event {
-                    ty: ev.ty.clone(),
-                    attributes: vec![],
-                };
-
-                for attr in &ev.attributes {
-                    if !filtered_attributes.contains(&attr.key) {
-                        new_ev.attributes.push(attr.clone());
-                    }
-                }
-
-                if new_ev.attributes.len() > 0 {
-                    events.push(new_ev);
-                }
-            }
-
-            SubMsgResult::Ok(SubMsgResponse {
-                events,
-                data: r.data.clone(),
-            })
-        }
-        SubMsgResult::Err(_) => reply.result.clone(),
-    };
-}
-
-// Parse the message that was passed to handle (Based on the assumption that it might be a reply or IBC as well)
-pub fn parse_message(
-    message: &[u8],
-    sig_info: &SigInfo,
-    handle_type: &HandleType,
-) -> Result<ParsedMessage, EnclaveError> {
-    let orig_secret_msg = SecretMessage::from_slice(message)?;
-
-    return match handle_type {
-        HandleType::HANDLE_TYPE_EXECUTE => {
-            trace!(
-                "handle input before decryption: {:?}",
-                base64::encode(&message)
-            );
-            let decrypted_msg = orig_secret_msg.decrypt()?;
-            Ok(ParsedMessage {
-                should_validate_sig_info: true,
-                was_msg_encrypted: true,
-                secret_msg: orig_secret_msg,
-                decrypted_msg,
-                contract_hash_for_validation: None,
-            })
-        }
-
-        HandleType::HANDLE_TYPE_REPLY => {
-            if sig_info.sign_mode == SignMode::SIGN_MODE_UNSPECIFIED {
-                trace!("reply input is not encrypted");
-                let decrypted_msg = orig_secret_msg.msg.clone();
-                let mut reply: Reply = serde_json::from_slice(&decrypted_msg)
-                    .map_err(|err| {
-                        warn!(
-                "reply got an error while trying to deserialize decrypted reply bytes into json {:?}: {}",
-                String::from_utf8_lossy(&decrypted_msg),
-                err
-            );
-                        EnclaveError::FailedToDeserialize
-                    })?;
-
-                let msg_id = String::from_utf8(reply.id.as_slice().to_vec()).map_err(|err| {
-                    warn!(
-                        "Failed to parse message id as string {:?}: {}",
-                        reply.id.as_slice().to_vec(),
-                        err
-                    );
-                    EnclaveError::FailedToDeserialize
-                })?;
-
-                let msg_id_as_num = match msg_id.parse::<u64>() {
-                    Ok(m) => m,
-                    Err(err) => {
-                        warn!("Failed to parse message id as number {}: {}", msg_id, err);
-                        return Err(EnclaveError::FailedToDeserialize);
-                    }
-                };
-
-                let decrypted_reply = DecryptedReply {
-                    id: msg_id_as_num,
-                    result: reply.result.clone(),
-                };
-
-                redact_custom_events(&mut reply);
-                let serialized_encrypted_reply : Vec<u8> = serde_json::to_vec(&reply).map_err(|err| {
-                    warn!(
-                        "got an error while trying to serialize encrypted reply into bytes {:?}: {}",
-                        reply, err
-                    );
-                    EnclaveError::FailedToSerialize
-                })?;
-
-                let reply_secret_msg = SecretMessage {
-                    nonce: orig_secret_msg.nonce,
-                    user_public_key: orig_secret_msg.user_public_key,
-                    msg: serialized_encrypted_reply,
-                };
-
-                let serialized_reply: Vec<u8> = serde_json::to_vec(&decrypted_reply).map_err(|err| {
-                    warn!(
-                        "got an error while trying to serialize decrypted reply into bytes {:?}: {}",
-                        decrypted_reply, err
-                    );
-                    EnclaveError::FailedToSerialize
-                })?;
-
-                return Ok(ParsedMessage {
-                    should_validate_sig_info: false,
-                    was_msg_encrypted: false,
-                    secret_msg: reply_secret_msg,
-                    decrypted_msg: serialized_reply,
-                    contract_hash_for_validation: None,
-                });
-            }
-
-            // Here we are sure the reply is OK because only OK is encrypted
-            trace!(
-                "reply input before decryption: {:?}",
-                base64::encode(&message)
-            );
-            let mut parsed_encrypted_reply: Reply = serde_json::from_slice(
-                &orig_secret_msg.msg.as_slice().to_vec(),
-            )
-            .map_err(|err| {
-                warn!(
-            "reply got an error while trying to deserialize msg input bytes into json {:?}: {}",
-            String::from_utf8_lossy(&orig_secret_msg.msg),
-            err
-            );
-                EnclaveError::FailedToDeserialize
-            })?;
-
-            match parsed_encrypted_reply.result.clone() {
-                SubMsgResult::Ok(response) => {
-                    let decrypted_msg_data = match response.data {
-                        Some(data) => {
-                            let tmp_secret_msg_data = SecretMessage {
-                                nonce: orig_secret_msg.nonce,
-                                user_public_key: orig_secret_msg.user_public_key,
-                                msg: data.as_slice().to_vec(),
-                            };
-
-                            Some(Binary(
-                                tmp_secret_msg_data.decrypt()?[HEX_ENCODED_HASH_SIZE..].to_vec(),
-                            ))
-                        }
-                        None => None,
-                    };
-
-                    let tmp_secret_msg_id = SecretMessage {
-                        nonce: orig_secret_msg.nonce,
-                        user_public_key: orig_secret_msg.user_public_key,
-                        msg: parsed_encrypted_reply.id.as_slice().to_vec(),
-                    };
-
-                    let tmp_decrypted_msg_id = tmp_secret_msg_id.decrypt()?;
-
-                    // Now we need to create synthetic SecretMessage to fit the API in "handle"
-                    let result = SubMsgResult::Ok(SubMsgResponse {
-                        events: response.events,
-                        data: decrypted_msg_data,
-                    });
-
-                    let msg_id =
-                        String::from_utf8(tmp_decrypted_msg_id[HEX_ENCODED_HASH_SIZE..].to_vec())
-                            .map_err(|err| {
-                            warn!(
-                                "Failed to parse message id as string {:?}: {}",
-                                tmp_decrypted_msg_id[HEX_ENCODED_HASH_SIZE..].to_vec(),
-                                err
-                            );
-                            EnclaveError::FailedToDeserialize
-                        })?;
-
-                    let msg_id_as_num = match msg_id.parse::<u64>() {
-                        Ok(m) => m,
-                        Err(err) => {
-                            warn!("Failed to parse message id as number {}: {}", msg_id, err);
-                            return Err(EnclaveError::FailedToDeserialize);
-                        }
-                    };
-
-                    let decrypted_reply = DecryptedReply {
-                        id: msg_id_as_num,
-                        result,
-                    };
-
-                    let decrypted_reply_as_vec =
-                        serde_json::to_vec(&decrypted_reply).map_err(|err| {
-                            warn!(
-                                "got an error while trying to serialize reply into bytes {:?}: {}",
-                                decrypted_reply, err
-                            );
-                            EnclaveError::FailedToSerialize
-                        })?;
-
-                    redact_custom_events(&mut parsed_encrypted_reply);
-                    let serialized_encrypted_reply : Vec<u8> = serde_json::to_vec(&parsed_encrypted_reply).map_err(|err| {
-                    warn!(
-                        "got an error while trying to serialize encrypted reply into bytes {:?}: {}",
-                        parsed_encrypted_reply, err
-                    );
-                    EnclaveError::FailedToSerialize
-                })?;
-
-                    let reply_secret_msg = SecretMessage {
-                        nonce: orig_secret_msg.nonce,
-                        user_public_key: orig_secret_msg.user_public_key,
-                        msg: serialized_encrypted_reply,
-                    };
-
-                    Ok(ParsedMessage {
-                        should_validate_sig_info: true,
-                        was_msg_encrypted: true,
-                        secret_msg: reply_secret_msg,
-                        decrypted_msg: decrypted_reply_as_vec,
-                        contract_hash_for_validation: Some(
-                            tmp_decrypted_msg_id[..HEX_ENCODED_HASH_SIZE].to_vec(),
-                        ),
-                    })
-                }
-                SubMsgResult::Err(response) => {
-                    let secret_msg = SecretMessage {
-                        nonce: orig_secret_msg.nonce,
-                        user_public_key: orig_secret_msg.user_public_key,
-                        msg: base64::decode(response.clone()).map_err(|err| {
-                            warn!(
-                                "got an error while trying to serialize err reply from base64 {:?}: {}",
-                                    response, err
-                            );
-                            EnclaveError::FailedToSerialize
-                        })?
-                    };
-
-                    let decrypted_error = secret_msg.decrypt()?;
-
-                    let tmp_secret_msg_id = SecretMessage {
-                        nonce: orig_secret_msg.nonce,
-                        user_public_key: orig_secret_msg.user_public_key,
-                        msg: parsed_encrypted_reply.id.as_slice().to_vec(),
-                    };
-
-                    let tmp_decrypted_msg_id = tmp_secret_msg_id.decrypt()?;
-
-                    // Now we need to create synthetic SecretMessage to fit the API in "handle"
-                    let result = SubMsgResult::Err(
-                        String::from_utf8(decrypted_error[HEX_ENCODED_HASH_SIZE..].to_vec())
-                            .map_err(|err| {
-                                warn!(
-                                    "Failed to parse error as string {:?}: {}",
-                                    decrypted_error[HEX_ENCODED_HASH_SIZE..].to_vec(),
-                                    err
-                                );
-                                EnclaveError::FailedToDeserialize
-                            })?,
-                    );
-
-                    let msg_id =
-                        String::from_utf8(tmp_decrypted_msg_id[HEX_ENCODED_HASH_SIZE..].to_vec())
-                            .map_err(|err| {
-                            warn!(
-                                "Failed to parse message id as string {:?}: {}",
-                                tmp_decrypted_msg_id[HEX_ENCODED_HASH_SIZE..].to_vec(),
-                                err
-                            );
-                            EnclaveError::FailedToDeserialize
-                        })?;
-
-                    let msg_id_as_num = match msg_id.parse::<u64>() {
-                        Ok(m) => m,
-                        Err(err) => {
-                            warn!("Failed to parse message id as number {}: {}", msg_id, err);
-                            return Err(EnclaveError::FailedToDeserialize);
-                        }
-                    };
-
-                    let decrypted_reply = DecryptedReply {
-                        id: msg_id_as_num,
-                        result,
-                    };
-
-                    let decrypted_reply_as_vec =
-                        serde_json::to_vec(&decrypted_reply).map_err(|err| {
-                            warn!(
-                                "got an error while trying to serialize reply into bytes {:?}: {}",
-                                decrypted_reply, err
-                            );
-                            EnclaveError::FailedToSerialize
-                        })?;
-
-                    let serialized_encrypted_reply : Vec<u8> = serde_json::to_vec(&parsed_encrypted_reply).map_err(|err| {
-                    warn!(
-                        "got an error while trying to serialize encrypted reply into bytes {:?}: {}",
-                        parsed_encrypted_reply, err
-                    );
-                    EnclaveError::FailedToSerialize
-                })?;
-
-                    let reply_secret_msg = SecretMessage {
-                        nonce: orig_secret_msg.nonce,
-                        user_public_key: orig_secret_msg.user_public_key,
-                        msg: serialized_encrypted_reply,
-                    };
-
-                    Ok(ParsedMessage {
-                        should_validate_sig_info: true,
-                        was_msg_encrypted: true,
-                        secret_msg: reply_secret_msg,
-                        decrypted_msg: decrypted_reply_as_vec,
-                        contract_hash_for_validation: Some(
-                            tmp_decrypted_msg_id[..HEX_ENCODED_HASH_SIZE].to_vec(),
-                        ),
-                    })
-                }
-            }
-        }
-    };
+fn to_canonical(contract_address: &BaseAddr) -> Result<CanonicalAddr, EnclaveError> {
+    CanonicalAddr::from_human(contract_address).map_err(|err| {
+        warn!(
+            "error while trying to deserialize address from bech32 string to bytes {:?}: {}",
+            contract_address, err
+        );
+        EnclaveError::FailedToDeserialize
+    })
 }
 
 #[cfg_attr(feature = "cargo-clippy", allow(clippy::too_many_arguments))]
@@ -532,50 +162,24 @@ pub fn handle(
     sig_info: &[u8],
     handle_type: u8,
 ) -> Result<HandleSuccess, EnclaveError> {
+    trace!("Starting handle");
+
     let contract_code = ContractCode::new(contract);
+    let contract_hash = contract_code.hash();
 
-    let mut env_v010: cw_types_v010::types::Env = serde_json::from_slice(env).map_err(|err| {
-        warn!(
-            "handle got an error while trying to deserialize env input bytes into json {:?}: {}",
-            env, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
-    env_v010.contract_code_hash = hex::encode(contract_code.hash());
+    let base_env: BaseEnv = extract_base_env(env)?;
 
-    trace!("handle env_v010: {:?}", env_v010);
+    let (sender, contract_address, _, sent_funds) = base_env.get_verification_params();
 
-    let canonical_contract_address = CanonicalAddr::from_human(&env_v010.contract.address).map_err(|err| {
-        warn!(
-            "got an error while trying to deserialize env_v010.contract.address from bech32 string to bytes {:?}: {}",
-            env_v010.contract.address, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let canonical_contract_address = to_canonical(contract_address)?;
 
-    let canonical_sender_address = CanonicalAddr::from_human(&env_v010.message.sender).map_err(|err| {
-        warn!(
-            "init got an error while trying to deserialize env_v010.message.sender from bech32 string to bytes {:?}: {}",
-            env_v010.message.sender, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let canonical_sender_address = to_canonical(sender)?;
 
-    let contract_key = extract_contract_key(&env_v010)?;
+    let contract_key = base_env.get_contract_key()?;
 
-    if !validate_contract_key(&contract_key, &canonical_contract_address, &contract_code) {
-        warn!("got an error while trying to deserialize output bytes");
-        return Err(EnclaveError::FailedContractAuthentication);
-    }
+    validate_contract_key(&contract_key, &canonical_contract_address, &contract_code)?;
 
-    let parsed_sig_info: SigInfo = serde_json::from_slice(sig_info).map_err(|err| {
-        warn!(
-            "handle got an error while trying to deserialize sig info input bytes into json {:?}: {}",
-            String::from_utf8_lossy(&sig_info),
-            err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let parsed_sig_info: SigInfo = extract_sig_info(sig_info)?;
 
     // The flow of handle is now used for multiple messages (such ash Handle, Reply)
     // When the message is handle, we expect it always to be encrypted while in Reply for example it might be plaintext
@@ -597,9 +201,9 @@ pub fn handle(
         // Verify env parameters against the signed tx
         verify_params(
             &parsed_sig_info,
-            &env_v010.message.sent_funds,
+            sent_funds,
             &canonical_sender_address,
-            &env_v010.contract.address,
+            contract_address,
             &secret_msg,
         )?;
     }
@@ -607,23 +211,10 @@ pub fn handle(
     let mut validated_msg = decrypted_msg.clone();
     let mut reply_params: Option<ReplyParams> = None;
     if was_msg_encrypted {
-        let x = validate_msg(
-            &decrypted_msg,
-            contract_code.hash(),
-            contract_hash_for_validation,
-        )?;
+        let x = validate_msg(&decrypted_msg, &contract_hash, contract_hash_for_validation)?;
         validated_msg = x.validated_msg;
         reply_params = x.reply_params;
     }
-
-    trace!(
-        "handle input afer decryption: {:?}",
-        String::from_utf8_lossy(&validated_msg)
-    );
-
-    trace!("Successfully authenticated the contract!");
-
-    trace!("Handle: Contract Key: {:?}", hex::encode(contract_key));
 
     // Although the operation here is not always handle it is irrelevant in this case
     // because it only helps to decide whether to check floating points or not
@@ -638,11 +229,15 @@ pub fn handle(
         secret_msg.user_public_key,
     )?;
 
-    let (contract_env_bytes, contract_msg_info_bytes) =
-        env_to_env_msg_info_bytes(&engine, &mut env_v010)?;
+    let mut versioned_env =
+        base_env.into_versioned_env(&engine.contract_instance.cosmwasm_api_version);
 
-    let env_ptr = engine.write_to_memory(&contract_env_bytes)?;
-    let msg_info_ptr = engine.write_to_memory(&contract_msg_info_bytes)?;
+    versioned_env.set_contract_hash(&contract_hash);
+
+    let (env_bytes, msg_info_bytes) = versioned_env.get_wasm_ptrs()?;
+
+    let env_ptr = engine.write_to_memory(&env_bytes)?;
+    let msg_info_ptr = engine.write_to_memory(&msg_info_bytes)?;
     let msg_ptr = engine.write_to_memory(&validated_msg)?;
 
     // This wrapper is used to coalesce all errors in this block to one object
@@ -661,7 +256,7 @@ pub fn handle(
             output,
             &secret_msg,
             &canonical_contract_address,
-            &env_v010.contract_code_hash,
+            versioned_env.get_contract_hash(),
             reply_params,
             &canonical_sender_address,
             false,
@@ -677,6 +272,17 @@ pub fn handle(
     Ok(HandleSuccess { output })
 }
 
+fn extract_sig_info(sig_info: &[u8]) -> Result<SigInfo, EnclaveError> {
+    serde_json::from_slice(sig_info).map_err(|err| {
+        warn!(
+            "handle got an error while trying to deserialize sig info input bytes into json {:?}: {}",
+            String::from_utf8_lossy(&sig_info),
+            err
+        );
+        EnclaveError::FailedToDeserialize
+    })
+}
+
 pub fn query(
     context: Ctx,
     gas_limit: u64,
@@ -685,46 +291,26 @@ pub fn query(
     env: &[u8],
     msg: &[u8],
 ) -> Result<QuerySuccess, EnclaveError> {
+    trace!("Entered query");
+
     let contract_code = ContractCode::new(contract);
+    let contract_hash = contract_code.hash();
 
-    let mut env_v010: cw_types_v010::types::Env = serde_json::from_slice(env).map_err(|err| {
-        warn!(
-            "query got an error while trying to deserialize env input bytes into json {:?}: {}",
-            env, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
-    env_v010.contract_code_hash = hex::encode(contract_code.hash());
+    let base_env: BaseEnv = extract_base_env(env)?;
 
-    trace!("query env_v010: {:?}", env_v010);
+    let (_, contract_address, _, _) = base_env.get_verification_params();
 
-    let canonical_contract_address = CanonicalAddr::from_human(&env_v010.contract.address).map_err(|err| {
-        warn!(
-            "got an error while trying to deserialize env_v010.contract.address from bech32 string to bytes {:?}: {}",
-            env_v010.contract.address, err
-        );
-        EnclaveError::FailedToDeserialize
-    })?;
+    let canonical_contract_address = to_canonical(contract_address)?;
 
-    let contract_key = extract_contract_key(&env_v010)?;
+    let contract_key = base_env.get_contract_key()?;
 
-    if !validate_contract_key(&contract_key, &canonical_contract_address, &contract_code) {
-        warn!("query got an error while trying to validate contract key");
-        return Err(EnclaveError::FailedContractAuthentication);
-    }
+    validate_contract_key(&contract_key, &canonical_contract_address, &contract_code)?;
 
-    trace!("successfully authenticated the contract!");
-    trace!("query contract key: {:?}", hex::encode(contract_key));
-
-    trace!("query input before decryption: {:?}", base64::encode(&msg));
     let secret_msg = SecretMessage::from_slice(msg)?;
     let decrypted_msg = secret_msg.decrypt()?;
-    trace!(
-        "query input afer decryption: {:?}",
-        String::from_utf8_lossy(&decrypted_msg)
-    );
+
     let ValidatedMessage { validated_msg, .. } =
-        validate_msg(&decrypted_msg, contract_code.hash(), None)?;
+        validate_msg(&decrypted_msg, &contract_hash, None)?;
 
     let mut engine = start_engine(
         context,
@@ -736,10 +322,14 @@ pub fn query(
         secret_msg.user_public_key,
     )?;
 
-    let (contract_env_bytes, _ /* no msg_info in query */) =
-        env_to_env_msg_info_bytes(&engine, &mut env_v010)?;
+    let mut versioned_env =
+        base_env.into_versioned_env(&engine.contract_instance.cosmwasm_api_version);
 
-    let env_ptr = engine.write_to_memory(&contract_env_bytes)?;
+    versioned_env.set_contract_hash(&contract_hash);
+
+    let (env_bytes, _) = versioned_env.get_wasm_ptrs()?;
+
+    let env_ptr = engine.write_to_memory(&env_bytes)?;
     let msg_ptr = engine.write_to_memory(&validated_msg)?;
 
     // This wrapper is used to coalesce all errors in this block to one object
@@ -797,77 +387,18 @@ fn start_engine(
     Ok(Engine::new(contract_instance, module))
 }
 
-fn env_to_env_msg_info_bytes(
-    engine: &Engine,
-    env_v010: &mut cw_types_v010::types::Env,
-) -> Result<(Vec<u8>, Vec<u8>), EnclaveError> {
-    match engine.contract_instance.cosmwasm_api_version {
-        CosmWasmApiVersion::V010 => {
-            // Assaf: contract_key is irrelevant inside the contract,
-            // but existing v0.10 contracts might expect it to be populated :facepalm:,
-            // therefore we are going to leave it populated :shrug:.
-            // env_v010.contract_key = None;
-
-            // in v0.10 the timestamp passed from Go was unix time in seconds
-            // 10.16 time is unix time in nanoseconds, so now Go passes here unix time in nanoseconds
-            // but v0.10 contracts still expect time to be in unix seconds,
-            // so we need to convert it from nanoseconds to seconds
-            env_v010.block.time = Timestamp::from_nanos(env_v010.block.time).seconds();
-
-            let env_v010_bytes = serde_json::to_vec(env_v010).map_err(|err| {
-                warn!(
-                    "got an error while trying to serialize env_v010 (cosmwasm v0.10) into bytes {:?}: {}",
-                    env_v010, err
-                );
-                EnclaveError::FailedToSerialize
-            })?;
-
-            let msg_info_v010_bytes: Vec<u8> = vec![]; // in v0.10 msg_info is inside env
-
-            Ok((env_v010_bytes, msg_info_v010_bytes))
-        }
-        CosmWasmApiVersion::V1 => {
-            let env_v1 = cw_types_v1::types::Env {
-                block: cw_types_v1::types::BlockInfo {
-                    height: env_v010.block.height,
-                    time: Timestamp::from_nanos(env_v010.block.time),
-                    chain_id: env_v010.block.chain_id.clone(),
-                },
-                contract: cw_types_v1::types::ContractInfo {
-                    address: Addr(env_v010.contract.address.0.clone()),
-                    code_hash: env_v010.contract_code_hash.clone(),
-                },
-            };
-
-            let env_v1_bytes =  serde_json::to_vec(&env_v1).map_err(|err| {
-                warn!(
-                    "got an error while trying to serialize env_v1 (CosmWasm v1) into bytes {:?}: {}",
-                    env_v1, err
-                );
-                EnclaveError::FailedToSerialize
-            })?;
-
-            let msg_info_v1 = cw_types_v1::types::MessageInfo {
-                sender: Addr(env_v010.message.sender.0.clone()),
-                funds: env_v010
-                    .message
-                    .sent_funds
-                    .iter()
-                    .map(|coin| {
-                        cw_types_v1::coins::Coin::new(coin.amount.u128(), coin.denom.clone())
-                    })
-                    .collect::<Vec<cw_types_v1::coins::Coin>>(),
-            };
-
-            let msg_info_v1_bytes =  serde_json::to_vec(&msg_info_v1).map_err(|err| {
-                warn!(
-                    "got an error while trying to serialize msg_info_v1 (CosmWasm v1) into bytes {:?}: {}",
-                    msg_info_v1, err
-                );
-                EnclaveError::FailedToSerialize
-            })?;
-
-            Ok((env_v1_bytes, msg_info_v1_bytes))
-        }
-    }
+fn extract_base_env(env: &[u8]) -> Result<BaseEnv, EnclaveError> {
+    serde_json::from_slice(env)
+        .map_err(|err| {
+            warn!(
+                "error while deserializing env into json {:?}: {}",
+                String::from_utf8_lossy(&env),
+                err
+            );
+            EnclaveError::FailedToDeserialize
+        })
+        .map(|base_env| {
+            trace!("base env: {:?}", base_env);
+            base_env
+        })
 }
