+++ cosmwasm/enclaves/shared/cosmwasm-types/generic/Cargo.toml
@@ -18,3 +18,4 @@ cw_types_v1 = { path = "../v1.0" }
 
 log = "0.4.14"
 hex = "0.4.2"
+base64 = "0.13.0"
