+++ cosmwasm/enclaves/shared/cosmwasm-types/v1.0/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw_types_v1"
-version = "1.2.4"
+version = "1.4.0"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
