+++ cosmwasm/enclaves/shared/crypto/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "enclave_crypto"
-version = "1.2.4"
+version = "1.4.0"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
