+++ cosmwasm/enclaves/shared/utils/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "enclave_utils"
-version = "1.2.4"
+version = "1.4.0"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
