+++ x/compute/internal/keeper/secret_contracts_test.go
@@ -610,12 +610,12 @@ func TestCallbackSanity(t *testing.T) {
 						{Key: "banana", Value: "🍌"},
 					},
 					{
-						{Key: "contract_address", Value: contractAddress.String()},
 						{Key: "kiwi", Value: "🥝"},
+						{Key: "contract_address", Value: contractAddress.String()},
 					},
 					{
-						{Key: "contract_address", Value: contractAddress.String()},
 						{Key: "watermelon", Value: "🍉"},
+						{Key: "contract_address", Value: contractAddress.String()},
 					},
 				},
 				execEvents,
